Move WriteConfig to its own module outside `element`
