+++ src/lazy/encoder/binary/v1_0/writer.rs
@@ -1,4 +1,3 @@
-use crate::element::writer::WriteConfigKind;
 use crate::lazy::encoder::binary::v1_0::value_writer::BinaryValueWriter_1_0;
 use crate::lazy::encoder::private::Sealed;
 use crate::lazy::encoder::value_writer::internal::MakeValueWriter;
@@ -7,7 +6,8 @@ use crate::lazy::encoder::write_as_ion::WriteAsIon;
 use crate::lazy::encoder::LazyRawWriter;
 use crate::lazy::encoding::Encoding;
 use crate::unsafe_helpers::{mut_ref_to_ptr, ptr_to_mut_ref};
-use crate::{IonResult, WriteConfig};
+use crate::write_config::{WriteConfig, WriteConfigKind};
+use crate::IonResult;
 use bumpalo::collections::Vec as BumpVec;
 use bumpalo::Bump as BumpAllocator;
 use delegate::delegate;
