+++ src/write_config.rs
@@ -1,19 +1,8 @@
-// Copyright Amazon.com, Inc. or its affiliates.
-
-//! Provides utility to serialize Ion data from [`Element`] into common targets
-//! such as byte buffers or files.
-
-use crate::result::IonResult;
-use crate::{Element, TextKind, Value};
-
-use crate::lazy::encoding::BinaryEncoding_1_1;
-#[cfg(feature = "experimental-lazy-reader")]
-use {
-    crate::lazy::encoder::{LazyEncoder, LazyRawWriter},
-    crate::lazy::encoding::{BinaryEncoding_1_0, Encoding, TextEncoding_1_0},
-    std::io,
-    std::marker::PhantomData,
-};
+use crate::lazy::encoder::{LazyEncoder, LazyRawWriter};
+use crate::lazy::encoding::{BinaryEncoding_1_0, BinaryEncoding_1_1, Encoding, TextEncoding_1_0};
+use crate::{IonResult, TextKind};
+use std::io;
+use std::marker::PhantomData;
 
 /// Writer configuration to provide format and Ion version details to writer through encoding
 /// This will be used to create a writer without specifying which writer methods to use
@@ -86,28 +75,3 @@ pub(crate) struct TextWriteConfig {
 // TODO: Add appropriate binary configuration if required for 1.1
 #[derive(Clone, Debug)]
 pub(crate) struct BinaryWriteConfig;
-
-/// Serializes [`Element`] instances into some kind of output sink.
-pub trait ElementWriter {
-    /// Serializes a single [`Value`] at the current depth of the writer.
-    fn write_value(&mut self, value: &Value) -> IonResult<()>;
-
-    /// Serializes a single [`Element`] at the current depth of the writer.
-    fn write_element(&mut self, element: &Element) -> IonResult<()>;
-
-    /// Serializes a collection of [`Element`].
-    ///
-    /// Most commonly used to serialize a series of top-level values, but can be used to write
-    /// [`Element`]s to an Ion `list` or `sexp` as well.
-    ///
-    /// This will return [`Err`] if writing any element causes a failure.
-    fn write_elements<'a, I: IntoIterator<Item = &'a Element>>(
-        &'a mut self,
-        elements: I,
-    ) -> IonResult<()> {
-        for element in elements.into_iter() {
-            self.write_element(element)?;
-        }
-        Ok(())
-    }
-}
